refactor(services): Add member library
