+++ services/apps/projects/src/project/project.controller.ts
@@ -3,8 +3,8 @@ import {NotFound, ObjectIdPipe} from '@mean-stream/nestx';
 import {Body, Controller, Delete, Get, Param, Patch, Post} from '@nestjs/common';
 import {ApiCreatedResponse, ApiOkResponse, ApiTags} from '@nestjs/swagger';
 import {Types} from 'mongoose';
-import {MemberAuth} from '../member/member-auth.decorator';
-import {MemberService} from '../member/member.service';
+import {MemberAuth} from '@app/member/member-auth.decorator';
+import {MemberService} from '@app/member/member.service';
 import {ProjectAuth} from './project-auth.decorator';
 import {CreateProjectDto, UpdateProjectDto} from './project.dto';
 import {Project} from './project.schema';
