+++ services/libs/member/src/index.ts
@@ -1,0 +1,5 @@
+export * from './member.dto';
+export * from './member.schema';
+export * from './member.service';
+export * from './member-auth.guard';
+export * from './member-auth.decorator';
