+++ services/nest-cli.json
@@ -69,6 +69,15 @@
       "compilerOptions": {
         "tsConfigPath": "libs/moss/tsconfig.lib.json"
       }
+    },
+    "member": {
+      "type": "library",
+      "root": "libs/member",
+      "entryFile": "index",
+      "sourceRoot": "libs/member/src",
+      "compilerOptions": {
+        "tsConfigPath": "libs/member/tsconfig.lib.json"
+      }
     }
   }
 }
