+++ services/package.json
@@ -121,7 +121,8 @@
     "moduleNameMapper": {
       "^@app/keycloak-auth(|/.*)$": "<rootDir>/libs/keycloak-auth/src/$1",
       "^@app/event(|/.*)$": "<rootDir>/libs/event/src/$1",
-      "^@app/moss(|/.*)$": "<rootDir>/libs/moss/src/$1"
+      "^@app/moss(|/.*)$": "<rootDir>/libs/moss/src/$1",
+      "^@app/member(|/.*)$": "<rootDir>/libs/member/src/$1"
     }
   },
   "pnpm": {
